Split bodyController off of KinematicBaseData
